Finished plyfile.spec to be VIP compliant

+++ external/src/SPECS/plyfile.spec
@@ -1,3 +1,5 @@
+%include %{_sourcedir}/common.inc
+
 %define name plyfile
 %define version 0.4
 %define unmangled_version 0.4
@@ -9,10 +11,10 @@ Name: %{name}
 Version: %{version}
 Release: %{release}
 Source0: %{name}-%{unmangled_version}.tar.gz
+Source1: common.inc
 License: GPLv3
 Group: Development/Libraries
 BuildRoot: %{_tmppath}/%{name}-%{version}-%{release}-buildroot
-Prefix: %{_prefix}
 BuildArch: noarch
 Vendor: Darsh Ranjan <darsh.ranjan@here.com>
 Url: https://github.com/dranjan/python-plyfile
@@ -24,13 +26,15 @@ Python module, which provides a simple facility for reading and writing ASCII an
 %setup -n %{name}-%{unmangled_version} -n %{name}-%{unmangled_version}
 
 %build
-python setup.py build
+env CFLAGS="$RPM_OPT_FLAGS" %{__python} setup.py build
 
 %install
-python setup.py install --single-version-externally-managed -O1 --root=$RPM_BUILD_ROOT --record=INSTALLED_FILES
+[ "$RPM_BUILD_ROOT" != "/" ] && rm -rf $RPM_BUILD_ROOT
+%{__python} setup.py install -O1 --prefix=${RPM_BUILD_ROOT}%{cat_prefix} --install-data=${RPM_BUILD_ROOT}%{_datadir} --single-version-externally-managed --record=/dev/null
 
 %clean
-rm -rf $RPM_BUILD_ROOT
+[ "$RPM_BUILD_ROOT" != "/" ] && rm -rf $RPM_BUILD_ROOT
 
-%files -f INSTALLED_FILES
+%files
 %defattr(-,root,root)
+/
